+++ include/ion-bb-dnn/edgetpu_c.h
@@ -2,7 +2,7 @@
 #define ION_BB_DNN_EDGETPU_C_H
 
 #include "tensorflowlite_c.h"
-#include "util.h"
+#include "rt_util.h"
 
 #ifdef __cplusplus
 extern "C" {
