+++ include/ion-bb-dnn/rt_util.h
